sugar: move many functions into new #aggregate import
